refactor: batch processor 위치 useSocketStore로 이동
