+++ client/src/stores/useSocketStore.ts
@@ -1,4 +1,3 @@
-import { BatchProcessor } from "@noctaCrdt/BatchProcessor";
 import {
   RemotePageCreateOperation,
   RemoteBlockInsertOperation,
@@ -15,6 +14,40 @@ import {
 } from "@noctaCrdt/Interfaces";
 import { io, Socket } from "socket.io-client";
 import { create } from "zustand";
+
+class BatchProcessor {
+  private batch: any[] = [];
+  private batchTimeout: number;
+  private batchTimer: any;
+  private sendBatchCallback: (batch: any[]) => void;
+
+  constructor(sendBatchCallback: (batch: any[]) => void, batchTimeout: number = 500) {
+    this.sendBatchCallback = sendBatchCallback;
+    this.batchTimeout = batchTimeout;
+  }
+
+  addOperation(operation: any) {
+    this.batch.push(operation);
+    if (!this.batchTimer) {
+      this.startBatchTimer();
+    }
+  }
+
+  private startBatchTimer() {
+    this.batchTimer = setTimeout(() => {
+      this.executeBatch();
+    }, this.batchTimeout);
+  }
+
+  private executeBatch() {
+    if (this.batch.length > 0) {
+      this.sendBatchCallback(this.batch);
+      this.batch = [];
+    }
+    clearTimeout(this.batchTimer);
+    this.batchTimer = null;
+  }
+}
 
 interface SocketStore {
   socket: Socket | null;
@@ -91,17 +124,14 @@ export const useSocketStore = create<SocketStore>((set, get) => ({
     });
 
     socket.on("assign/clientId", (clientId: number) => {
-      console.log("Assigned client ID:", clientId);
       set({ clientId });
     });
 
     socket.on("workspace", (workspace: WorkSpaceSerializedProps) => {
-      console.log("Received initial workspace state:", workspace);
       set({ workspace });
     });
 
     socket.on("connect", () => {
-      console.log("Connected to server");
       set({ socket });
     });
